# src/backend/docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Car Share API
  description: This is the API for the Car Share project.
  version: 0.4.0 # TODO: Update version number
  contact:
    name: Car Share, backend team
    email: 'contoux1u@etu.univ-lorraine.fr'
  license:
    name: GNU General Public License v3.0
    url: 'https://www.gnu.org/licenses/gpl-3.0.en.html'
servers:
  - url: 'https://carshare.api.antoinectx.fr'
    description: Antoine's server
  - url: 'https://carshare.apis.furwaz.fr'
    description: Paul's server
...
paths:
...
  /groups:
    get:
      summary: Search groups (admin)
      description: "Search in all groups. Only available to admins.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: searchGroups
      tags:
        - Groups
        - Admin
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/query'
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: A list of the groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
                  query:
                    type: string
                    required: false
                    example: friends
                  limit:
                    type: integer
                    example: 10
                  offset:
                    type: integer
                    example: 0
                  next:
                    type: integer
                    nullable: true
                    example: 10
                  prev:
                    type: integer
                    nullable: true
                    example: 0
                  total:
                    type: integer
                    example: 12
        '401':
          $ref: '#/components/responses/noToken'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
  /groups/my:
    post:
      summary: Create a group
      description: "Create a group.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: createGroup
      tags:
        - Groups
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
      security:
        - bearerAuth: [ ]
      requestBody:
        description: The group to create
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  required: true
                  example: My friends
      responses:
        '200':
          description: Group created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Group created
                  group:
                    $ref: '#/components/schemas/Group'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Group name is required.
        '401':
          $ref: '#/components/responses/noToken'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
    get:
      summary: Get my groups
      description: Get my groups.
      operationId: getMyGroups
      tags:
        - Groups
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Groups retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
                  query:
                    type: string
                    required: false
                    example: friends
                  limit:
                    type: integer
                    example: 10
                  offset:
                    type: integer
                    example: 0
                  next:
                    type: integer
                    nullable: true
                    example: 10
                  prev:
                    type: integer
                    nullable: true
                    example: 0
                  total:
                    type: integer
                    example: 12
        '401':
          $ref: '#/components/responses/noToken'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
  /groups/:id/member:
    post:
      summary: Add a new member
      description: "Add a new member in a group.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: addUserGroup
      tags:
        - Groups
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: id
          in: path
          description: The id of the group to which a user is added
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: [ ]
      requestBody:
        description: The email of the user to add on the group
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  required: true
                  example: john@example.com
      responses:
        '200':
          description: User added
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User added
                  group:
                    $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/invalidId'
        '401':
          $ref: '#/components/responses/noToken'
        '403':
          $ref: '#/components/responses/groupNotCreator'
        '404':
          $ref: '#/components/responses/groupNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
    delete:
      summary: Remove a member
      description: "Remove a member of a group.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: removeUserGroup
      tags:
        - Groups
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: id
          in: path
          description: The id of the group to which a user is removed
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: [ ]
      requestBody:
        description: The email of the user to add on the group
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  required: true
                  example: john@example.com
      responses:
        '200':
          description: Member removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Member removed
        '400':
          $ref: '#/components/responses/invalidId'
        '401':
          $ref: '#/components/responses/noToken'
        '403':
          $ref: '#/components/responses/groupNotCreator'
        '404':
          $ref: '#/components/responses/groupNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
  /groups/:id:
    delete:
      summary: Delete a group (admin)
      description: "Delete a group, with the id in parameter. Only available to admins.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: deleteGroup
      tags:
        - Groups
        - Admin
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: id
          in: path
          description: The id of the group to delete
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Group removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Group removed
        '400':
          $ref: '#/components/responses/invalidId'
        '401':
          $ref: '#/components/responses/noToken'
        '403':
          $ref: '#/components/responses/insufficientPrivileges'
        '404':
          $ref: '#/components/responses/groupNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
    patch:
      summary: Update a group (admin)
      description: "Update a group.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: updateGroup
      tags:
        - Groups
        - Admin
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: groupeId
          in: path
          description: The id of the group to update
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                groupName:
                  type: string
                  example: Bilboquet
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Group updated
        '400':
          $ref: '#/components/responses/invalidId'
        '401':
          $ref: '#/components/responses/noToken'
        '403':
          $ref: '#/components/responses/insufficientPrivileges'
        '404':
          $ref: '#/components/responses/groupNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
  /groups/my/:id:
    delete:
      summary: Delete a group
      description: "Delete a group created by the user, with the id in parameter.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: deleteMyGroup
      tags:
        - Groups
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: id
          in: path
          description: The id of the group to delete
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Group removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Group removed
        '400':
          $ref: '#/components/responses/invalidId'
        '401':
          $ref: '#/components/responses/noToken'
        '403':
          $ref: '#/components/responses/groupNotCreator'
        '404':
          $ref: '#/components/responses/groupNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
    patch:
      summary: Update a group
      description: "Update a group owned by the current user.
      <br /><strong>A verified email is required to perform this action.</strong>"
      operationId: updateMyGroup
      tags:
        - Groups
      parameters:
        - $ref: '#/components/parameters/Accept-Language'
        - name: groupeId
          in: path
          description: The id of the group to update
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                groupName:
                  type: string
                  example: Bilboquet
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Group updated
        '400':
          $ref: '#/components/responses/invalidId'
        '401':
          $ref: '#/components/responses/noToken'
        '403':
          $ref: '#/components/responses/groupNotCreator'
        '404':
          $ref: '#/components/responses/groupNotFound'
        '498':
          $ref: '#/components/responses/invalidToken'
        '500':
          $ref: '#/components/responses/internalError'
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  parameters:
    Accept-Language:
      name: Accept-Language
      in: header
      description: The language to use in the response message (if any)
      required: false
      schema:
        type: string
        default: en
        enum:
          - en
          - fr
    limit:
      name: limit
      in: query
      description: The number of items to return
      required: false
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 50
    offset:
      name: offset
      in: query
      description: The number of items to skip before starting to collect the result set
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    query:
      name: query
      in: query
      description: The query to search for
      required: false
      schema:
        type: string
        default: ''
  responses:
    groupNotCreator:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: You are not the creator of this group.
    invalidId:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Id is invalid.
    noToken:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: No token provided.
    invalidToken:
      description: Invalid/Expired token
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Invalid token provided.
    insufficientPrivileges:
      description: Insufficient privileges
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: You do not have sufficient rights to perform this action.
...
    groupNotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Group not found.
...
    internalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: An internal error occurred. If the problem persists, please contact an administrator.
...
  schemas:
...
    Group:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: My friends
        creatorId:
          type: integer
          example: 1
        creator:
          $ref: '#/components/schemas/PublicUser'
        createdAt:
          type: string
          format: date-time
          example: 2020-01-01T00:00:00.000Z
        users:
          type: array
          items:
            $ref: '#/components/schemas/PublicUser'
...
    PublicUser:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: john@example.com
        emailVerifiedOn:
          type: string
          format: date-time
          nullable: true
          default: null
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        phone:
          type: string
          example: "+330612345678"
        avatar:
          type: string
          nullable: true
          default: null
        gender:
          type: integer
          default: -1
        hasCar:
          type: boolean
          default: false
        level:
          type: integer
          default: 0
        lang:
          type: string
          nullable: true
          default: null
          example: fr
        timezone:
          type: string
          nullable: true
          default: null
          example: Europe/Paris
...